Change texi2pdf call to include local TeX configuration file which increases save_history variable.
The current pdf documentation cannot be built unless the save_size variable is increased from
its value of 5000.  On UNIX systems, not all users will have local root and the ability to
change the default.  I work around this by defining a local configuration file to use
in addition to the system-wide texmf.cnf file.