locale support implementation in [fs]scanf functions
Implement the ability to scan "1,2" and get 1.2 as results under
French locale, where the decimal separator is ','. This uses the C++
standard library locale approach: the interpretation of some stream
locale is modified using the "imbue" function. In case of fscanf, the
previous locale is restored after each call. The functions interface
is not modified. When the last argument is a string, its value is used
as the new locale setting.

* src/oct-stream.h (octave_base_stream::imbue): new function, which
merely ignores its argument. (octave_stream::imbue): add a real
implementation, which is a proxy to std::[io]stream.imbue.

* src/file_io.cc: (fscanf): when the last argument is a string, save
the actual locale value and set a new value from this string. Generate
a warning in case the requested locale is not supported. Revert to
previous locale if OK, or to standard locale in case of problem.
(scanf): Likewise, except the stream is generated on-the-fly, there is
no need to restore its locale value. Added a test case. Updated the
doc for both functions.