10/10 commits reworking the lexer
This and the previous nine changesets completely revise the way
significant whitespace is handled inside matrix and cell lists.  Now
instead of indefinite look ahead, we keep track of all tokens read and
look back to previous tokens and ahead at most one character.  If the
surrounding context changes the parse, we unput the text for the
current token and alter the start state, unput an additional character
(typically a comma or semicolon) or both.

Comments are now handled by a separate set of exlusive start states
and there is no explicit handling of comments within any other
patterns.