Overhaul max/min implementations for Sparse matrices to fix bug #40743.
* CSparse.cc (max, min): Declare variables for matrix A, rows and columns, and
B, rows and columns.  Test A & B sizes and if there is a mismatch check that
one of them is not a null matrix before issuing an error.

* dSparse.cc (max, min): Declare variables for matrix A, rows and columns, and
B, rows and columns.  Test A & B sizes and if there is a mismatch check that
one of them is not a null matrix before issuing an error.