bitcmp: fix bitwise complement for signed integers.
* general/bitcmp.m: current implementation of bitwise complement is broken
for signed integers because it uses intmax() as mask for bitxor().  This is
incorrect because the last bit of a signed integer is 0 and not 1.  Fix this
by bitpacking an array of true().  Also simplified input checking by using
sizeof, rather than multiple calls to isa().  Added tests for signed integers.