fieldnames: fix for java objects from dynamic classpath (bug #42710)
* fieldnames.m: this function supports passing a java class name as input
instead of a java object.  org.octave.ClassHelper.getFields() is capable
of accepting both a java object or a java class name.  However, it would
get confused if the object is a java string.  I will guess this was the
reason to always pass the java class name to it.  However, it seems that
it will only accept java class names on the static classpath, which made
the whole function failing when passing objects.  The function will still
fail if input is the classname of a class in the dynamic classpath though,
but at least will now work if an object is used.