Do not rebuild MIME database during install (backout 737b5d449aaf) (bug #44404)
* etc/icons/Makefile.am: the desktop file being installed have MIME type
information which means that the MIME types application database must be
rebuilt after install.  However, this means that we also need to take care of
it during uninstall and this is causing 'make distcheck' to fail (bug #44404).
The freedesktop.org recommendation seems to be that this should be left to
downstream packagers and that it is reasonable to expect that users know about
rebuilding this database themselves (just like ldconfig).  So do a partial
backout of 737b5d449aaf, leave the MIMEType line and leave the update of MIME
cache up to the user.