Automatically convert octave vectors into java primitive arrays (bug #44882)
* libinterp/octave-value/ov-java.cc (unbox): conversion was working fine when
input were scalars.  However, when input is a vector, Matlab automatically
converts into a java array of the corresponding primitive types.  This seems
to lose some information since a uint8 will be casted into a int8 for example.
The documentation is not super clear, maybe the casting needs to be done based
on the type of the method being called?