Backout rev af5591ef9790.
* scripts/plot/util/__gnuplot_drawnow__.m: When printing, the paper properties
are modified to size all objects to 1 pixel/point. This is needed to support
the design of GL2PS. This is also done for Gnuplot to avoid toolkit specific
code in print.m and __print_parse_opts__.m. Toolkit specific code shoudl be
limited to toolkit specific m-files, i.e.__opengl_print__.m, and
__gnuplot_print__.m. The changeset being backedout is associated with bug
report #46122.