imformats: return empty instead of error if there's no support for image IO.
* __magick_read__.cc (__magick_formats__): if Octave was built without
support for image IO, return an empty struct instead of error. If we
have no support for image format, it doesn't matter if that's because
of limited configuration on the GraphicsMagick build or Octave.
Remember we can't even promise support for all formats at the same time,
that's dependent on the configuration of GraphicsMagick.
* imformats.m: add simple test for even when we were built without
support for image IO.