pkg: support PKG_ADD file in inst/.
* pkg/private/create_pkgadddel.m: inst/ is meant to be the directory of
the package "ready" to install" and all files within it are copied to
the arch independent directory of the package.  However, when reading
the m files for PKG_ADD directives, it clobers an existing PKG_ADD file.
So append to such file if it exists.
This change will break "pkg build" since it is unable to guess the right
place for the PKG_ADD (but it is already broken, see bug #45369).