pkg: add support for [a-z.+-~] characters on package version number string.
* pkg/private/get_description.m: currently pkg enforces a x.y.z (major,
minor, and patch) numeric style for the string representing a package
version.  In addition, it will silently append a ".0" if the string is
missing the patch version which causes weird errors for downstream
packagers.  However, there seems to be no reason in the code for this
limitation which prevents 'x.y.z+' or 'x.y.z-rcN'.  A real limitation we
will have is that version string will be used in the package filepath.
This change just expands the valid characters for the version string.
It does not change the requirement when installing packages with the
-forge option though (but Octave forge packages only have the x.y.z
version string anyway).