acinclude.m4: Use pkg-config from OCTAVE_CHECK_LIB whenever possible
The following patch uses pkg-config to add extra flags if pkg-config
finds the module in question in OCTAVE_CHECK_LIB. This was necessary
for me in Debian in order to find the HDF5 libraries, and seems useful
in general.

Like it often happens with m4 and autoconf, I cargo-culted and
generalised this code from existing checks for sndfile in
configure.ac. It probably would make sense to move the original code
into its own m4 macro, so we can have a little less code duplication
in the autoconf files.

* acinclude.m4 (OCTAVE_CHECK_LIB): add a generic pkg-config check.