svd: reduce code duplication with more use of template and macro.
* liboctave/numeric/svd.cc, liboctave/numeric/svd.h: remove unused
constructor with reference for int (info).  This allows to move all
of the constructor into a single template, so remove init().  Two
new methods, gesvd and gesdd, are fully specialized but the main
hunck of code are the long list of arguments.  Scope type and drive
enums to the svd class for clarity, and rename member names.  Add
a new member for the drive used.
* libinterp/corefcn/svd.cc: fix typenames for the svd enums which
are now scoped.
* CMatrix.cc, dMatrix.cc, fCMatrix.cc, fMatrix.cc: fix typenames
for the svd enums which are now scoped.