Don't mark nested functions in primary function as subfunctions.
* help.cc (Flocalfunctions): Remove note from docstring about nested functions
causing problems with localfunctions usage.  Add @seealso to functions.
Don't check that function is a nested one before adding it to list.  If it is
in the subfunction_names list, it is valid.

* ov-fcn-handle.cc (octave_fcn_handle::octave_fcn_handle): In constructor,
only error out if function is a true nested function, rather than a subfunction
which happens to be the parent of a nested function.

* oct-parse.in.yy (base_parser::finish_function): Determine whether function
is a subfunction or nested function and only if it is a subfunction is it added
to the list of subfunctions of the primary function and marked as being a
subfunction.