fix caching error with Qt tool configuration (bug #50129)
* acinclude.m4 (OCTAVE_CHECK_QT_VERSION): Use separate variable names
when checking for Qt tools wit and without versions appended.  For
example, use MOC_QTVER when checking for "moc-qt$ver" and MOC when
checking for "moc".  Likewise for other Qt tools (uic, rcc, lrelease).
If all tools with the version number appended are found, set the simple
variable to the value of the version variable (MOC from MOC_QTVER, for
example).