check if input is class method before declaring it unimplemented (patch #9238) (bug #49694)
Octave helpfully checks for functions that are implemented in Octave Forge
packages.  If one of those function names is actually a class method (of a
class provided by the package) and the package is loaded, then the resulting
warning incorrectly states it is unimplemented.  Instead, we give a hint and
link to the appropriate help call.  This commit adds this support for the
"symbolic" package but the support is general if other packages need it too.

* scripts/help/__unimplemented__.m: improve behaviour for class methods

Pushed by siko1056 <k.ohlhus@gmail.com>.