Replace static instances of OCTAVE_LOCAL_BUFFER with local variables.
* load-save.cc (check_gzip_magic): Replace OCTAVE_LOCAL_BUFFER magic with
"unsigned char magic[2]".

* regexp.cc (octregexp): Replace OCTAVE_LOCAL_BUFFER_INIT arg_used with
"int arg_used[7] {}".  Aggregate initialization zero initializes ints.

* ccolamd.cc (Fccolamd, Fcsymand): Replace OCTAVE_LOCAL_BUFFER instances for
knobs and stats with 1) a local variable XXX_storage, and 2) a pointer to
the first element in XXX_storage.  Also use static_assert to guarantee
at compile time that the size of the local variable storage is not too large.

* colamd.cc (Fcolamd, Fsymand): Replace OCTAVE_LOCAL_BUFFER instances for
knobs and stats with 1) a local variable XXX_storage, and 2) a pointer to
the first element in XXX_storage.  Also use static_assert to guarantee
at compile time that the size of the local variable storage is not too large.

* Array.cc (hermitian): Remove comment about OCTAVE_LOCAL_BUFFER not being
appropriate.