Anonymous functions to local functions in classdef files (bug #51599).
* libinterp/corefcn/symtab.cc (symbol_table::fcn_info::fcn_info_rep::xfind):
When looking up local functions and no function file name is present, which is
the case for an anonymous function, it now looks at the parent scope for it.
Another approach of stashing the file information for an anonymous function
during the initialization in tree_evaluator::visit_anon_fcn_handle caused weird
bugs.

* test/bug-51599/bug-51599.tst, test/bug-51599/class_bug51599.m,
test/bug-51599/module.mk: new files for simple test of calling a local function
of a classdef class from an instance of that class using a function handle.

* test/module.mk: Add bug-51599 tests to build system.

Pushed by K. T. Ohlhus <k.ohlhus@gmail.com>.