publish: improve detection of inline markup and fix MathJax usage.
* scripts/general/publish.m: improve detection of inline markup by using more
sophisticated regular expressions compatible to Matlab.  Let the formatter
decide what to do with inline and block math, as it is not for granted, that
all possible output formats accept "$" and "$$" as math delimiters.

* scripts/general/private/__publish_latex_output__.m: New formatting functions
for inline and block math.  Improved documentation.

* scripts/general/private/__publish_html_output__.m: New formatting functions
for inline and block math.  Improved documentation.  Use another CDN according
to shutdown of MathJaxCDN https://www.mathjax.org/cdn-shutting-down/.